fix the namespace and refactor RectanglePlace
